+++ src/components/UsageLimitsCard/components/UsageLimits/UsageLimits.scss
@@ -1,4 +1,4 @@
-@import '../../styles/common';
+@import '../../../../styles/common';
 
 .TotalUsageLimitTextField {
   @include constrained-input;
