+++ src/components/UsageLimitsCard/components/UsageLimits/index.ts
@@ -1,0 +1,1 @@
+export {UsageLimits} from './UsageLimits';
