+++ src/components/UsageLimitsCard/components/index.ts
@@ -1,1 +1,2 @@
 export {RecurringPayment} from './RecurringPayment';
+export {UsageLimits} from './UsageLimits';
